# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "VEGAS POINT APIs LIST",
      "description": "docs for vegas point *****",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "192.168.100.57:8090/api",
    "basePath": "/",
...
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
      "/point/{id}": {
        "get": {
          "tags": ["Points"],
          "parameters": [
            {
              "name": "id",
              "default": 1,
              "in": "path",
              "required": true,
              "description": "ID of user that we want to find",
              "type": "integer"
             
            }
          ],
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "get list of point by id successfull"
            }
          }
        }
      },
      "/point_by_date_cardtrack": {
        "post": {
          "tags": ["Points"],
          "parameters": [
            {
              "name": "point_model",
              "in": "body",
              "required": true,
              "description": "find all point by date range with card\nUSE '/card_number/{id}' to get id",
              "schema": {
                "$ref": "#/definitions/Point"
              }
  
            }
          ],
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "get list of point by id successfull"
            }
          }
        }
      },
      "/point_by_date_cardtrack/range": {
        "post": {
          "tags": ["Points"],
          "parameters": [
            {
              "name": "pointrange_model",
              "in": "body",
              "required": true,
              "description": "find all point by date range and id\nUSER '/card_number/{id}' to get ID",
              "schema": {
                "$ref": "#/definitions/PointRange"
              }
            }
          ],
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "get list of point by id successfull"
            }
          }
        }
      },
      "/card_number/{id}": {
        "get": {
          "tags": ["Cards"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "default": 1,
              "required": true,
              "description": "ID of user that we want to find cardID - cardID use for find point by cardtrack",
              "type": "integer"
            }
          ],
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "get list of card successfull"
            }
          }
        }
      },
      "/machine_player": {
        "summary": "OK",
        "post": {
          "tags": ["Machines"],
          "description": "machines player find by id",
          "parameters": [
            {
              "name": "machine_model",
              "in": "body",
              "description": "machine_model_body",
              "schema": {
                "$ref": "#/definitions/Machine"
              }
            }
          ],
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "get data successfull"
            }
          }
        }
      },
      "/jackpot_history": {
        "summary": "OK",
        "post": {
          "tags": ["Jackpots"],
          "description": "find jackpot by date: start - end",
          "parameters": [
            {
              "name": "jackpot_model",
              "in": "body",
              "description": "jackpot_model_body",
              "schema": {
                "$ref": "#/definitions/Jackpot"
              }
            }
          ],
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "get data successfull"
            }
          }
        }
      }
    },
    "definitions": {
      "Machine": {
        "required": ["customer_number", "date"],
        "properties": {
          "customer_number": {
            "type": "integer",
            "default": 1
          },
          "date": {
            "type": "string",
            "default": "2022-08-08"
          }
        }
      },
      "Jackpot": {
        "required": ["startDate","endDate"],
        "properties": {
          "startDate": {
            "type": "string",
            "default": "2022-08-08"
          },
          "endDate": {
            "type": "string",
            "default": "2022-08-09"
          }
        }
      },
      "Point": {
        "required": ["dateToday","dateToday2","startDateWeek","endDateWeek","startDateMonth","endDateMonth","id"],
        "properties": {
          "id": {
            "type": "string",
            "default": "1003005455"
          },
          "dateToday": {
            "type": "string",
            "default": "2022-07-04"
          },
          "dateToday2": {
            "type": "string",
            "default": "2022-07-05"
          },
          "startDateWeek": {
            "type": "string",
            "default": "2022-07-04"
          },
          "endDateWeek": {
            "type": "string",
            "default": "2022-07-10"
          },
          "startDateMonth": {
            "type": "string",
            "default": "2022-07-01"
          },
          "endDateMonth": {
            "type": "string",
            "default": "2022-07-31"
          }
        }
      },
      "PointRange": {
        "required": ["startDate","endDate","id"],
        "properties": {
          "id": {
            "type": "string",
            "default": "1003005455"
          },
          "startDate": {
            "type": "string",
            "default": "2022-08-08"
          },
          "endDate": {
            "type": "string",
            "default": "2022-08-09"
          }
        }
      }
    }
  }